Make FHIR Client Socket Timeout Configurable
CQl queries are evaluated using synchronous HTTP requests. The HAPI FHIR
client has a default socket timeout of 10 seconds. This will be not
sufficient for longer running queries. I added the Spring config option
`app.store.socketTimeout` in order to be able to increase the timeout
with a default of 60 seconds.

I added the integration test DktkFedSearchApplicationSlowQueryTest which
uses an nginx container to slow down requests to 2 each minute, so that
each request to Blaze takes 30 seconds.